Extract ConfigureKernel trait
